Remove a couple of TODOs
So far, I've been able to match anything I may want to match using the
app name. It may not always be 100% as accurate (especially if the
pattern is very short and ends up being matched unvoluntarily, but I
will only do something more complicated if experience indicates that
it's actually a problem)

I will also add more logging if I need it, but not for the sake of it.