feat: Ability to use icons for empty workspace.
This allows users of ex. `waybar` to use dynamic icons,
while still giving empty workspaces an icon instead of a number.
This is implemented by renaming new workspaces right when they
are created.
This is all behind a config option; defaults remain unchanged.

Implementing this is also an optimization for the rest of the program,
which no longer recalculates workspace naming on _every_ `Event::Window`.
Instead, only specific `Window` (and now also `Workspace`) events
will do so.

**Caveats**:
- There is a small visual artifact when opening a new workspace,
when the number gets renamed to the desired icon.. Unfortunately,
being a program reacting to sway events, I don't see a way to fix this.

**Future Possibilities**:
- Allowing specific icons for specific numbered (empty) workspaces
wouldn't be difficult to add (which would close #33).
- Which `*.change` types precisely should trigger workspace renaming
was determined with common sense, and could do with a one-over.
The relevant function is `should_rename_after_event`.

Closes #39.